requirements.txt
pyautogui==0.9.54
pytesseract==0.3.10
Pillow==10.4.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1
schedule==1.2.2
opencv-python==4.10.0
numpy==1.26.4

# macOS specific dependencies
pyobjc-core>=9.0
pyobjc-framework-Quartz>=9.0
